Replace React/Flask with Streamlit implementation
- Simplified architecture to single Streamlit app
- Removed React frontend and Flask API backend
- Maintained same functionality for marginal child benefit analysis
- Kept PolicyEngine branding and colors
- Simplified deployment process
