+++ README.md
@@ -1,108 +1,83 @@
 # The Marginal Child
 
-An interactive web application that analyzes how government benefits change with each additional child, powered by [PolicyEngine-US](https://github.com/PolicyEngine/policyengine-us).
+A Streamlit application that analyzes how government benefits change with each additional child, powered by PolicyEngine.
+
+## Overview
+
+This application visualizes the marginal benefit of having additional children across different income levels, taking into account various government programs including:
+
+- SNAP (Food Stamps)
+- WIC (Women, Infants, and Children nutrition program)
+- Medicaid/CHIP (Healthcare)
+- Premium Tax Credits
+- EITC (Earned Income Tax Credit)
+- Child Tax Credit
+- Child and Dependent Care Credit
 
 ## Features
 
-- **Dynamic Benefit Calculations**: Uses PolicyEngine-US for accurate, up-to-date calculations of:
-  - SNAP (Food Stamps)
-  - WIC
-  - Medicaid
-  - CHIP (Children's Health Insurance Program)
-  - Premium Tax Credits (ACA Marketplace)
-  - EITC (Earned Income Tax Credit)
-  - Child Tax Credit
-  - Child and Dependent Care Credit
-  - Housing subsidies
-
-- **Interactive Visualizations**:
-  - Benefit cliff analysis showing how benefits phase out with income
-  - Net income vs employment income comparison
-  - Marginal benefit per additional child across income ranges
-
-- **State-Specific Analysis**: Calculate benefits for all 50 states plus DC
-
+- **Interactive Visualizations**: Real-time chart updates showing net income changes per additional child
+- **State-Specific Analysis**: Calculate benefits for all 50 US states plus DC
 - **Customizable Household Configuration**:
-  - Marital status
-  - Number and ages of children
-  - Housing and childcare costs
-  - Employment income for both spouses
+  - Marital status (single/married)
+  - State selection
+  - Spouse income (if married)
+- **Summary Statistics**: Average benefits for 1st, 2nd, and 3rd children
 
 ## Installation
 
 ### Prerequisites
 
-- Python 3.10+
-- Node.js 22+
-- uv (Python package manager)
+- Python 3.9+
 
 ### Setup
 
 1. Clone the repository:
 ```bash
-git clone https://github.com/yourusername/the-marginal-child.git
+git clone https://github.com/PolicyEngine/the-marginal-child.git
 cd the-marginal-child
 ```
 
-2. Install Python dependencies:
+2. Install dependencies:
 ```bash
-cd api
-uv venv
-uv pip install -r requirements.txt
-cd ..
-```
-
-3. Install Node dependencies:
-```bash
-npm install
+pip install -r requirements.txt
 ```
 
 ## Running the Application
 
-You'll need to run both the Python API backend and the React frontend:
-
-### Terminal 1: Start the API server
 ```bash
-cd api
-uv run python app.py
+streamlit run app.py
 ```
-The API will run on http://localhost:5000
 
-### Terminal 2: Start the React development server
-```bash
-npm run dev
-```
-The app will open at http://localhost:5173
-
-## Usage
-
-1. **Configure Household**: Set your household parameters including marital status, state, number of children, and income.
-
-2. **Add Child Ages**: Specify ages for each child (affects benefit eligibility).
-
-3. **Set Analysis Parameters**: Choose the income range and step size for the analysis.
-
-4. **Run Analysis**:
-   - **Calculate Current Benefits**: See benefits at your specified income level
-   - **Analyze Benefit Cliff**: Visualize how benefits change across income ranges
-   - **Analyze Marginal Child**: See the marginal benefit of each additional child
+The app will open in your browser at http://localhost:8501
 
 ## How It Works
 
-The application uses PolicyEngine-US's microsimulation model to calculate benefits based on actual federal and state policy rules. Unlike simple calculators with hard-coded values, this app:
+The application calculates the marginal benefit of each additional child by:
+1. Computing net income (earnings + benefits) for households with 0-4 children
+2. Calculating the difference in net income for each additional child
+3. Displaying these marginal benefits across the full income range ($0-$200,000)
 
-- Uses actual policy parameters (Federal Poverty Guidelines, benefit formulas, etc.) from PolicyEngine
-- Accounts for interactions between programs
-- Includes state-specific variations
-- Updates automatically as PolicyEngine updates its models
+All children are assumed to be age 10 for benefit calculation purposes.
 
 ## Technology Stack
 
-- **Backend**: Flask API with PolicyEngine-US
-- **Frontend**: React with Vite
-- **UI Components**: Material-UI
-- **Charts**: Plotly.js
-- **Calculations**: PolicyEngine-US microsimulation engine
+- **Framework**: Streamlit for rapid prototyping and deployment
+- **Visualization**: Plotly for interactive charts
+- **Data Processing**: Pandas and NumPy
+- **Styling**: PolicyEngine brand colors and design system
+
+## Deployment
+
+The app can be deployed to:
+- Streamlit Cloud (recommended for quick deployment)
+- Any cloud provider supporting Python web apps
+- Docker containers
+
+For Streamlit Cloud deployment:
+1. Push to GitHub
+2. Connect repository to Streamlit Cloud
+3. Deploy with one click
 
 ## Contributing
 
@@ -114,4 +89,4 @@ MIT
 
 ## Acknowledgments
 
-Built using [PolicyEngine-US](https://policyengine.org), the open-source microsimulation model for US tax and benefit policy.
+Powered by [PolicyEngine](https://policyengine.org), the open-source microsimulation infrastructure for tax and benefit policy.
