--- api/requirements.txt
@@ -1,5 +1,0 @@
-flask
-flask-cors
-policyengine-us>=1.0.0
-pandas
-numpy
