+++ requirements.txt
@@ -1,0 +1,4 @@
+streamlit>=1.28.0
+plotly>=5.17.0
+pandas>=2.1.0
+numpy>=1.24.0
